made examples into function
